登陆退出接口

+++ src/views/Body/Exit/Exit.vue
@@ -14,9 +14,9 @@
                     </div>
                 </div>
                 <div class="content_form">
-                   <p>
-                       确定要退出吗 ？
-                   </p>
+                    <p>
+                        确定要退出吗 ？
+                    </p>
                 </div>
                 <div class="footer_btn">
                     <el-button @click="cancel" size="small">取 消</el-button>
@@ -28,6 +28,8 @@
 </template>
 
 <script>
+    import Api from '../../../api/Login/Login'
+
     export default {
         props: {
             isExit: {
@@ -41,45 +43,54 @@
             },
             confirm() {
                 this.$emit("update:isExit", false);
-                window.localStorage.clear();
-                window.location.reload();
-                this.$router.push('/')
+                Api.outPost().then((res) => {
+                    if (res.code === '200') {
+                        window.localStorage.clear();
+                        window.location.reload();
+                        this.$router.push('/')
+                    }
+                })
+
             }
         },
     }
 </script>
 
 <style scoped>
-    .updPass_box{
+    .updPass_box {
         position: fixed;
         top: 0;
         width: 100%;
         height: 100%;
-        background: rgba(0,0,0,0.5);
+        background: rgba(0, 0, 0, 0.5);
         z-index: 99;
     }
-    .updPass{
+
+    .updPass {
         width: 100%;
         height: 100%;
         position: relative;
     }
-    .updPass_conntent{
+
+    .updPass_conntent {
         position: absolute;
         top: 50%;
         left: 50%;
-        transform: translate(-50%,-50%);
+        transform: translate(-50%, -50%);
         background: #ffffff;
         width: 300px;
         height: 180px;
         border-radius: 5px;
     }
-    .upd_title{
+
+    .upd_title {
         width: 100%;
         background: #eee;
         border-top-left-radius: 5px;
         border-top-right-radius: 5px;
     }
-    .title_box{
+
+    .title_box {
         padding: 0 20px;
         width: 90%;
         height: 40px;
@@ -87,28 +98,34 @@
         justify-content: space-between;
         align-items: center;
     }
-    .title{
+
+    .title {
         font-size: 16px;
     }
-    .el-icon-close{
+
+    .el-icon-close {
         font-size: 18px;
         font-weight: bold;
     }
-    .footer_btn{
+
+    .footer_btn {
         width: 100%;
         text-align: center;
         position: fixed;
         bottom: 0;
         padding: 20px 0;
     }
-    .content_form{
+
+    .content_form {
         width: 100%;
     }
-    .content_form>p{
-       margin-top: 30px;
+
+    .content_form > p {
+        margin-top: 30px;
         margin-left: 20px;
     }
-    .int_box>span{
-        padding-right:20px;
+
+    .int_box > span {
+        padding-right: 20px;
     }
 </style>
